Added a Framework
The Snowflake Framework is a maneuver I've recently implemented for my notes that combines the likes of Canvas and Markdown files. It creates a hybrid conceptual note, providing the best of both worlds!

Simply make a .md file, link a canvas file with !![[]], and link the markdown file with other files through bulletpoints listed as keywords.

Hope this helps!